Reorganize code to pass flake8 complexity test
Flake8 complexity tests failed due to overly complex methods in several
files.

This fixes that by creating simpeler methods. The
`async_message_broker.py` is suspected to be (maybe slightly modified)
copy of code, and could probably be extracted as shared code to
GOB-Utils or something like that. For now it has been annotated with
`#noqa`