Remove GOB-Core tests from testphase
Automatic testing in Docker failed, because GOB-Core is installed with
tests, and expanded in the /app/ dir. Automatic test-detection found
them and tried to run them.

This removes the test, and lets automatic test-detection only test the
GOB-Upload tests. This line can be removed once we no longer use github
checkouts but PyPI package.