Improve performance of comparison
A shallow compare is executed by calculating a hash of the new data and inserting it in a temporary table. This table is then compared to the current state which will return ADD, CONFIRM and DELETE events. The remaining message will be compared on an attribute level to create the MODIFY events. When a collection is inserted for the first time ADD events will be created without creating the temporary table