Fix the bug with auto id
On partial imports, some records could already have been assigned a value on a previous run. This value was being used to determine the next value, but it would not necessarily be the highest value in the database. To prevent this, we will set the last value when the enricher is initialized